Create DB conn and SQL Queries for testing the KingsDB in JAVA
When running this you have to change minor details after you run the first time. The ID's are primary keys and data integrity makes it so we cannot duplicate ID's/Primary key values. When you want to run it again just delete the rows that were added, or change the ID values to avoid duplicating primary key values. If you forget this eclipse will throw error messages and tell you the exact same thing.
I'll have to send you the DB files so that you'll have the exact same DB as me and the code will run smoothly.